Unit tests merged into PluginsGenerationServiceTests class
